requirements.txt
pandas
numpy
plotly
openai
pandasql
sentence-transformers
unidecode
google-generativeai